Corrected GfxKeyEvent ctor

+++ include/RE/G/GFxEvent.h
@@ -95,7 +95,7 @@ namespace RE
 			keyboardIndex(0)
 		{}
 
-		inline GFxKeyEvent(EventType a_eventType, GFxKey::Code a_keyCode, std::uint32_t a_ascii, std::uint32_t a_wchar, GFxSpecialKeysState a_specialKeyState, std::uint8_t a_keyboardIndex = 0) :
+		inline GFxKeyEvent(EventType a_eventType, GFxKey::Code a_keyCode, std::uint8_t a_ascii, std::uint32_t a_wchar, GFxSpecialKeysState a_specialKeyState, std::uint8_t a_keyboardIndex = 0) :
 			GFxEvent(a_eventType),
 			keyCode(a_keyCode),
 			asciiCode(a_ascii),
